requirements.txt
explainit-1.0-py3-none-any.whl
pandas==1.4.2
gunicorn==20.1.0
Werkzeug==2.0.0
explainit-1.0-py3-none-any.whl